pembenaran int ke float kamar

+++ application/controllers/administrasi/Tagihan.php
@@ -1135,10 +1135,10 @@ class Tagihan extends CI_Controller
                         $tanggal_cek_out_ri_kamar = $this->input->post('tanggal_cek_out_ri_kamar')[$i];
 
                         $jumlah_hari_temp = $this->input->post('jumlah_hari_ri_kamar')[$i];
-                        $jumlah_hari = (int) $jumlah_hari_temp;
+                        $jumlah_hari = (float) $jumlah_hari_temp;
 
                         $harga_harian_temp = $this->input->post('harga_harian_ri_kamar')[$i];
-                        $harga_harian = (int) preg_replace("/[^0-9]/", "", $harga_harian_temp);
+                        $harga_harian = (float) preg_replace("/[^0-9]/", "", $harga_harian_temp);
 
                         $status_kamar_ri_kamar = $this->input->post('status_kamar_ri_kamar')[$i];
 
@@ -1304,10 +1304,10 @@ class Tagihan extends CI_Controller
                         $tanggal_cek_out_ri_kamar = $this->input->post('tanggal_cek_out_ri_kamar')[$i];
 
                         $jumlah_hari_temp = $this->input->post('jumlah_hari_ri_kamar')[$i];
-                        $jumlah_hari = (int) $jumlah_hari_temp;
+                        $jumlah_hari = (float) $jumlah_hari_temp;
 
                         $harga_harian_temp = $this->input->post('harga_harian_ri_kamar')[$i];
-                        $harga_harian = (int) preg_replace("/[^0-9]/", "", $harga_harian_temp);
+                        $harga_harian = (float) preg_replace("/[^0-9]/", "", $harga_harian_temp);
 
                         $status_kamar_ri_kamar = $this->input->post('status_kamar_ri_kamar')[$i];
